Introduce cache flipping
We want to care for the possible bottleneck in the online alter binlog
reading/writing in advance.

IO_CACHE does not provide anything better that sequential access, besides,
only a single write is mutex-protected, which is not suitable, since we
should write a transaction atomically.

To solve this, a special layer on top Event_log is implemented.
There are two IO_CACHE files underneath: one for reading, and one for
writing.

Once the read cache is empty, an exclusive lock is acquired (we can wait
for a currently active transaction finish writing), and flip() is emitted,
i.e. the write cache is reopened for read, and the read cache is emptied,
and reopened for writing.

This reminds a buffer flip that happens in accelerated graphics
(DirectX/OpenGL/etc).

Cache_flip_event_log is considered non-blocking for a single reader and a
single writer in this sense, with the only lock held by reader during flip.

An alternative approach by implementing a fair concurrent circular buffer
is described in MDEV-24676.